fixed unnecessary compilation of SBML and Sundials in linux build script

+++ CompuCell3D/build_scripts/linux/build-debian-cc3d.sh
@@ -162,7 +162,7 @@ then
   mkdir -p $BUILD_ROOT/CompuCell3D
   cd $BUILD_ROOT/CompuCell3D
 
-  run_and_watch_status COMPUCELL3D_CMAKE_CONFIG cmake -G "Unix Makefiles" --build=/home/m/CompuCell3D_build -DCMAKE_INSTALL_PREFIX:PATH=$INSTALL_PREFIX $SOURCE_ROOT/CompuCell3D 
+  run_and_watch_status COMPUCELL3D_CMAKE_CONFIG cmake -G "Unix Makefiles" -DCMAKE_INSTALL_PREFIX:PATH=$INSTALL_PREFIX $SOURCE_ROOT/CompuCell3D 
   run_and_watch_status COMPUCELL3D_COMPILE_AND_INSTALL make $MAKE_MULTICORE_OPTION && make install
   
   ############# END OF BUILDING CC3D
@@ -176,28 +176,41 @@ then
   export CXXFLAGS=-fPIC
   export CFLAGS=-fPIC
 
+  SBML_BUILD_DIR=$BUILD_ROOT/libsbml-3.4.1
+  SBML_INSTALL_DIR=$DEPENDENCIES_ROOT/libsbml-3.4.1
 
-  cp $SOURCE_ROOT/BionetSolver/dependencies/libsbml-3.4.1-src.zip $BUILD_ROOT
-  cd $BUILD_ROOT 
-
-  unzip libsbml-3.4.1-src.zip
-
-  cd $BUILD_ROOT/libsbml-3.4.1
-  run_and_watch_status LIBSBML_CONFIGURE ./configure --prefix=$DEPENDENCIES_ROOT/libsbml-3.4.1
   
-  # libsbml does not compile well with multi-core option on
+  if [ ! -d "$SBML_INSTALL_DIR" ]; then # SBML_INSTALL_DIR does not exist
   
-  run_and_watch_status LIBSBML_COMPILE_AND_INSTALL make  && make install
+    cp $SOURCE_ROOT/BionetSolver/dependencies/libsbml-3.4.1-src.zip $BUILD_ROOT
+    cd $BUILD_ROOT 
 
-  cp $SOURCE_ROOT/BionetSolver/dependencies/sundials-2.3.0.tar.gz $BUILD_ROOT
-  cd $BUILD_ROOT
+    unzip libsbml-3.4.1-src.zip
+    
+    cd $SBML_BUILD_DIR 
+    run_and_watch_status LIBSBML_CONFIGURE ./configure --prefix=$SBML_INSTALL_DIR 
+    # libsbml does not compile well with multi-core option on
+    run_and_watch_status LIBSBML_COMPILE_AND_INSTALL make  && make install
+  fi
+  
+  
+  
+  
+  SUNDIALS_BUILD_DIR=$BUILD_ROOT/sundials-2.3.0
+  SUNDIALS_INSTALL_DIR=$DEPENDENCIES_ROOT/sundials-2.3.0
+  
+  if [ ! -d "$SUNDIALS_INSTALL_DIR" ]; then # SUNDIALS_INSTALL_DIR does not exist
+    
+    cp $SOURCE_ROOT/BionetSolver/dependencies/sundials-2.3.0.tar.gz $BUILD_ROOT
+    cd $BUILD_ROOT
 
-  tar -zxvf sundials-2.3.0.tar.gz
+    tar -zxvf sundials-2.3.0.tar.gz
 
-  cd $BUILD_ROOT/sundials-2.3.0
-  run_and_watch_status SUNDIALS_CONFIGURE ./configure --with-pic --prefix=$DEPENDENCIES_ROOT/sundials-2.3.0
+    cd SUNDIALS_BUILD_DIR
+    run_and_watch_status SUNDIALS_CONFIGURE ./configure --with-pic --prefix=$SUNDIALS_INSTALL_DIR
 
-  run_and_watch_status SUNDIALS_COMPILE_AND_INSTALL  make $MAKE_MULTICORE_OPTION && make install
+    run_and_watch_status SUNDIALS_COMPILE_AND_INSTALL  make $MAKE_MULTICORE_OPTION && make install
+  fi
   ############# END OF BUILDING SBML AND SUNDIALS BIONET DEPENDENCIES
 fi
